Added render queue submit

+++ src/util/Types.h
@@ -5,12 +5,12 @@
 typedef unsigned char U8;
 typedef unsigned short int U16;
 typedef unsigned int U32;
-typedef unsigned long int U64;
+typedef unsigned long long int U64;
 
 typedef char I8;
 typedef short int I16;
 typedef int I32;
-typedef long int I64;
+typedef long long int I64;
 
 typedef float F32;
 typedef double F64;
